Minor cleanups to eliminate compilation warnings.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@1085 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/winvnc/vncauth.c
@@ -57,7 +57,7 @@ vncEncryptPasswd(char *passwd, char *encryptedPasswd)
     /* pad password with nulls */
 
     for (i = 0; i < MAXPWLEN; i++) {
-	if (i < strlen(passwd)) {
+	if (i < (int)strlen(passwd)) {
 	    encryptedPasswd[i] = passwd[i];
 	} else {
 	    encryptedPasswd[i] = 0;
@@ -117,7 +117,7 @@ vncEncryptBytes(unsigned char *where, const char *passwd)
     /* key is simply password padded with nulls */
 
     for (i = 0; i < 8; i++) {
-	if (i < strlen(passwd)) {
+	if (i < (int)strlen(passwd)) {
 	    key[i] = passwd[i];
 	} else {
 	    key[i] = 0;
